requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
langchain
langchain-ollama
langchain-community
pydantic
pdfplumber
python-multipart